backend/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.2
django-cors-headers==4.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
pillow==11.0.0
psycopg==3.2.3
psycopg2==2.9.10
PyJWT==2.9.0
razorpay==1.4.2
requests==2.32.3
rest-framework-simplejwt==0.0.2
setuptools==75.3.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
